+++ docs/source/changelog.md
@@ -6,6 +6,7 @@
 - Add debug function `materialize_table`.
 - Update install instructions and dependencies to enable DB2 and mssql development on OS X with an `arm64` architecture.
 - Update PR template
+- Run `RUNSTATS` on every DB2 table after creation
 
 ## 0.6.6 (2023-08-17)
 - Implement support for loading polars dataframes from DuckDB.